requirements.txt
asgiref==3.7.2
Django==4.1.1
django-debug-toolbar==4.1.0
django-environ==0.10.0
django-filter==23.2
djangorestframework==3.14.0
coreapi==2.3.3
iniconfig==2.0.0
Markdown==3.4.3
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4